Shuffle namespace under capistrano.
