+++ lib/capistrano/gitflow/natcmp.rb
